requirements.txt
streamlit==1.44.1
pandas==2.2.3
pypdf==5.4.0
sentence-transformers==3.4.1
faiss-cpu==1.10.0
numpy==2.2.4
scikit-learn==1.6.1
python-dotenv==1.1.0
openai==1.74.0
requests==2.32.3